Fix viewport resize bug (#1880)
* fix: viewport resize bug
